Fix #36
The HTML looks like it has changed. I (think) all the URIs for each
song are hosted from "http://23.237.126.42/ost..." so instead of
scraping from the text "Click here to download as...", scrape from
the matching href attribute that starts with the URI starting as
listed above. Tested and seems to fix the issues. However not sure
if the songs may be hosted on another server, which would break
this fix.